Continuation on methodology
Continued integrating already existing parts into the new template.
Changed some stuff around with the `cleveref` library (as the apendix references did not work well, due to the document type).

I also chose to discuss the "weight adding" part of the system under "data enhancement" with a new subchapter, as it fits better there.
The data interpretation part will now take this as a given and focus on calculating angles, populating angle-ranges (corresponding to arm zones), and the interpretation of that data (including stabalization).